churn_agent/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.4.0
sentence-transformers>=2.7.0
google-generativeai>=0.7.0
groq>=0.9.0
python-dotenv>=1.0.0
